% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAGOSNEgis-package.R
\docType{package}
\name{LAGOSUSgis-package}
\alias{LAGOSUSgis}
\alias{LAGOSUSgis-package}
\title{LAGOSUSgis: Extra functions to interact with the GIS module of LAGOSUS}
\description{
Extra Functions to Supplement LAGOSUS.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jsta/LAGOSUSgis}
  \item Report bugs at \url{https://github.com/jsta/LAGOSUSgis/issues}
}

}
\author{
\strong{Maintainer}: Jemma Stachelek \email{stachel2@msu.edu}

}
\keyword{internal}
