% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lagosusgis_path}
\alias{lagosusgis_path}
\title{Get the path to LAGOSUSgis data}
\usage{
lagosusgis_path()
}
\description{
Get the path to LAGOSUSgis data
}
\examples{
\dontrun{
# use defaults
lagosusgis_path()

# custom setting
Sys.setenv(
LAGOSUSGIS_PATH = path.expand("~/.local/share/LAGOS-GIS/gis_locus_v1.0.gpkg"))
lagosusgis_path()

# revert to defaults
Sys.setenv(LAGOSUSGIS_PATH = "")
lagosusgis_path()
}
}
