% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_lake_poly}
\alias{query_lake_poly}
\title{Query polygon outline for LAGOS lakes}
\usage{
query_lake_poly(
  lagoslakeid,
  gis_path = lagosusgis_path(),
  crs = albers_conic(),
  utm = FALSE
)
}
\arguments{
\item{lagoslakeid}{numeric}

\item{gis_path}{file.path to LAGOSUS GIS gpkg}

\item{crs}{projection string or epsg code}

\item{utm}{logical convert crs to utm}
}
\description{
Query polygon outline for LAGOS lakes
}
\examples{
\dontrun{
res <- query_lake_poly(c(474805, 474803))
}
}
